Checks whether metadata is in scrattch.taxonomy format

Messages will be printed to the screen.

Usage

checkMetadata(
  meta.data,
  schema,
  messages = c(),
  isValid = FALSE,
  isWarning = FALSE,
  print.messages = TRUE,
  ...
)

Arguments

meta.data

A meta.data object to be tested

schema

The schema data.frame.

messages

The current messages to append to.

isValid

The current call for whether the taxonomy is valid.

isWarning

The current call for whether the taxonomy is valid.

print.messages

Print messages only to a log file (FALSE; default) or also to the screen (TRUE)

...

Additional parameters for ._validate_schema_element and .validate_var_elements (can be ignored in most cases)

Value

Logical vector indicating whether the inputted taxonomy is a valid scrattch.taxonomy format.