Updates a metadata data frame with ontology terms

See updateTaxonomyMetadata for details. computeOntologyTerms is a wrapper for updateTaxonomyMetadata with defaults to do compute the ontology terms for everything (except CL), but not do anything else. Any breaking issues will cause this function to return FALSE. And potential issues will still return TRUE but will output a warning to stderr. All messages will get returned to the log file.

Usage

computeOntologyTerms(
  metadata,
  log.file.path = getwd(),
  log.file.name = "computeOntologyTerms_log.txt",
  standardize.metadata = FALSE,
  compute.ontology.terms = c("organism", "anatomical_region", "self_reported_sex",
    "self_reported_ethnicity", "assay", "disease"),
  compute.brain.atlas.terms = "DHBA",
  convert.regions.to.names = TRUE,
  compute.cl.terms = NULL,
  print.messages = FALSE
)

Value

A list where "metadata" is the updated metadata file and there are additional list entries corresponding to conversions surrounding ontology terms. The original metadata data frame is also returned.