This function will return the best matching ontology term id given a name and a distance to that best match

This function will return the best matching ontology term id given a name and a distance to that best match

Usage

._find_best_ontology_match(search_term, ontology, top_n = 1)

Arguments

search_term

Name of an ontology term you want to try and match (e.g., "frontal cortex" or "human")

ontology

Variable of class ontology_index that holds a complete ontology read in from obo format using get_OBO.

top_n

Number of top matches to return (default = 1)

Value

character indicating the associated id name and distance for the closest matching term