resolve_cl

resolve_cl

Usage

.resolve_cl(
  cl.g,
  cl.dat,
  markers,
  map.dat,
  select.cells,
  p = 0.8,
  low.th = 0.1,
  seed = 42
)

Arguments

cl.g

Cluster labels in some format

markers

Genes to use as markers for this function

map.dat

normalized data of the MAPPING data set.

select.cells

names of cells to map (e.g., the column names of the cell x gene matrix)

p

proportion of marker genes to include in each iteration of the mapping algorithm.

low.th

the minimum difference in Pearson correlation required to decide on which branch

seed

random seed for reproducibility

cl.med

Cluster medians

Value

mapped.cl output