Generates the precomputed_stats.h5 file, which is an HDF5 file that contains a serialization of the cell type taxonomy tree, as well as statistics about the reference cells that have been assigned to the cell types in the taxonomy.

Generates the precomputed_stats.h5 file, which is an HDF5 file that contains a serialization of the cell type taxonomy tree, as well as statistics about the reference cells that have been assigned to the cell types in the taxonomy.

Usage

run_precomp_stats(
  anndata_path,
  n_processors,
  normalization,
  tmp_dir,
  hierarchy
)

Arguments

anndata_path

Local file path of the AIT reference taxonomy (h5ad file).

n_processors

Number of independent worker processes to spin up.

normalization

Normalization of the h5ad files; must be either 'raw' or 'log2CPM'.

tmp_dir

Temporary directory for writing out the hierarchical files (the code will clean these up after itself).

hierarchy

List of term_set_labels in the reference taxonomy ordered from most gross to most fine.

Value

File path to the precompute stats file.