Saves the contents of the query_markers.json file to AIT -> uns -> hierarchical.

Saves the contents of the query_markers.json file to AIT -> uns -> hierarchical.

Usage

save_query_markers_to_uns(AIT_anndata, query_markers_output_path)

Arguments

AIT_anndata

AIT reference taxonomy object where the file contents will be saved.

query_markers_output_path

Local file path to the generated or user provided query_markers.json file.

Value

AIT reference taxonomy with the query markers saved in uns -> hierarchical.